refactor: separated `add()` from `handler.rs`
